CODA-19 - Add Cypress integration tests
