+++ package.json
@@ -18,7 +18,7 @@
     "csslint": "stylelint ./src/**/*.scss",
     "csslint:fix": "stylelint ./src/**/*.scss --fix",
     "flow": "flow",
-    "prettify": "prettier --write \"./src/**/*.{js,jsx}\"",
+    "prettify": "prettier --write \"./src/**/*.{js,jsx}\" && prettier --write \"./example-server/**/*.{js,jsx}\"",
     "cypress:open": "cypress open",
     "cypress:run": "cypress run",
     "example-run": "node ./example-server/index.js"
